requirements.txt
Django==5.1.4
psycopg2-binary==2.9.9
django-import-export==3.3.8
django-prometheus==2.3.1
prometheus-client==0.19.0
influxdb-client==1.44.0
reportlab==4.0.8
xlsxwriter==3.1.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
papaparse==0.2.0